backend/requirements.txt
# ===== Core Web Framework =====
# pydantic：FastAPI / pydantic-settings 的基础，显式约束大版本便于锁定行为
fastapi==0.135.1
pydantic==2.12.5
pydantic-settings==2.13.1
uvicorn==0.42.0
python-multipart==0.0.22
email-validator==2.3.0

# ===== Authentication =====
python-jose[cryptography]==3.5.0
bcrypt==5.0.0

# ===== Database =====
sqlalchemy==2.0.48
alembic==1.18.4
mysql-connector-python==9.6.0

# ===== Object Storage =====
minio==7.2.20

# ===== Document Processing =====
# Docx2txtLoader 依赖 docx2txt，不是 python-docx
docx2txt==0.9
pypdf==6.9.1
# unstructured 0.21.x 支持 Python 3.9~3.12；如使用 Python 3.13 需 0.16.19+
unstructured==0.21.5

# ===== LangChain Core =====
langchain==1.2.13
langchain-core==1.2.20
langchain-text-splitters==1.1.1
langchain-community==0.4.1
langchain-openai==1.1.11

# ===== Vector Stores =====
chromadb==1.5.5
langchain-chroma==1.1.0

# ===== LLM Providers =====
openai==2.29.0
langchain-ollama==1.0.1

# ===== RAG Evaluation (Optional) =====
ragas==0.4.3
datasets==4.8.3

# ===== Dev / tests =====
pytest==8.4.2
